Add memoization to the `clean_key` function for improved perf.
`clean_key` is rather expensive due to the use of a regex, but
encounters a high amount of duplicate inputs. It is a good candidate
for memoizing, since it is deterministic. Future improvements may
be possible by splitting the key in '.', and doing a memoized
`clean_key` on the sub-components.